Bug 563810 - [Passage] implement issued licenses report (headless part)
 - rework test environment for better smoke and failure testing
